+++ api/internal/controller/rest/v2/controller.go
@@ -3,9 +3,9 @@ package v2
 import (
 	"net/http"
 
+	"github.com/erwin-lovecraft/aegismiles/internal/core/service/customer"
+	"github.com/erwin-lovecraft/aegismiles/internal/core/service/mileage"
 	"github.com/erwin-lovecraft/aegismiles/internal/models/dto"
-	"github.com/erwin-lovecraft/aegismiles/internal/services/customer"
-	"github.com/erwin-lovecraft/aegismiles/internal/services/mileage"
 	"github.com/viebiz/lit"
 	"github.com/viebiz/lit/iam"
 )
