+++ api/internal/core/domain/domain.go
@@ -1,0 +1,10 @@
+package domain
+
+import "github.com/erwin-lovecraft/aegismiles/internal/entity"
+
+type (
+	Customer       = entity.Customer
+	AccrualRequest = entity.AccrualRequest
+	MilesLedger    = entity.MilesLedger
+	TravelDistance = entity.TravelDistance
+)
